refactor!: add network dependency in context
A dependency on the `Network` singleton was forgotten to be extracted into the explicit context. Add a `OffchainNetworkState` record to represent the minimally necessary data that the implementation must obtain from the outer world (what the `Network` instance was previously used for). Include that as a field in the Context.