refactor: make `TrieAccessor` sealed
We want to encapsulate the intention of `BlockTrieAccessor` and `ChildTrieAccessor` being the two variants of what a `TrieAccessor` might constitute.

BlockTrieAccessor is intentionally left as-is (seemingly obsolete; with no additional data or functionality) as a marker class to type-hint that a TrieAccessor must be exactly the accessor for the main trie for the block; and not a child trie accessor.