fix(FullSyncMachine)!: block body decode
A breaking bug was introduced in #e222d8 as a result of moving block en/decoding logic from `FullSyncMachine` into the `Runtime`'s private methods. Turns out, blocks were being wrongly decoded for a long time in the `FullSyncMachine` (and persisted with the wrong bodies in the BlockState, for that matter, so it was actually and wider and older bug).