fix: Fix issue with connectivity to new peers. (#805)
# Description

Our current implementation only relies on the initial connections to
boot nodes when communicating with the chain. That is not a scalable
solution and does not conform to the spec. This PR aims to look for new
peers and send them handshakes in a scheduled manner so that we can
establish connections with them. Kademlia's `bootstrap(Host us)` method
finds peers close to us and establishes a connection with them. However,
we must send handshakes to the newly opened connection, or they remain
idle most of the time.