feat: Implemented BeefyState (#804)
# Description
Implemented BeefyState class responsible for managing BEEFY finality
rounds.
**Added essential fields for tracking:**
- Validator sets per session.
- Mandatory and non-mandatory block numbers for efficient finality.
- BEEFY genesis, finalization state, and next digest tracking.

**Introduced BeefySession class to store:**
- Authority sets for each session.
- A synchronized set of non-mandatory blocks that can be finalized for
lower latency.

**P.S.**
Sessions are created by a validator set change in a block - we create a
key value from the blockNumber and a session.

Non-mandatory blocks are added when votes occur.
We find the first session whose mandatory block number is greater than
voted block and store that block in the previous to that session as a
non-mandatory block.

Fixes #777 >

---------

Co-authored-by: Hristiyan Mitov <hristiyan.mitov@limechain.tech>
Co-authored-by: Georgi Grigorov <72125420+Grigorov-Georgi@users.noreply.github.com>