feat: Initiate grandpa. (#820)
# Description

The PR implements the following logic related to GRANDPA:
- Initializing a round at the end of the full sync process.
- Appropriately start rounds on catch up responses.
- Starting next round from within a round when a certain criteria is
met.

Currently if we are not a part of the voter set we do not have have the
logic to apply incoming justifications. This means that in a set that we
are not voting in we would be stagnant in the first round. That is in no
way ideal and will be improved in the future.