feat: Implement handling of consensus messages. (#827)
# Description

- Implemented handling of consensus messages
**(BEEFY_CHANGED_AUTHORITIES, BEEFY_ON_DISABLED)**.
- Introduces method, named
**retrieveMandatoryBlockAndCreateBeefySession()** - it examines BEEFY
authority consensus messages, within a specified block range to detect
authority set changes or disabled authorities. This method is being used
in two places in Beefy:
1) in **handleBeefySessionTransitions()** - it is applied to all the
blocks between beefyDigest and grandpa finalized. Each found mandatory
block with its corresponding beefy session is appended. This method is
being used when we have to basically go through blocks that are between
**lastBeefyFinalized**/**beefyGenesis** and **grandpaFinalized**

2) in **processJustification()** - it is applied to each of the incoming
**justifications**(signed commitments) from our peers. This method will
be used to **trigger verifying of justification, checking/storing
justifications from/to DB, update beefy state etc.**
I provided partial implementation of this method - only parts that are
related to **creating** and **erasing** **sessions** and **blocks**.
Rest of the method is left with //TODOs for future implementation.
     


Fixes #773

---------

Co-authored-by: Hristiyan Mitov <hristiyan.mitov@limechain.tech>
Co-authored-by: Yordan Atanasov <yordan.atanasov@limechain.tech>