feat: handle vote and justification messages (#875)
# Description

- Introduced cachedAcceptedInterval filed for storing start & end round
numbers. Its purpose is to reduce the calls to acceptInterval() method.
CachedAcceptedInterval filed is being updated on finalization of a beefy
round and finalizedBlockChanged event.
- Implemented filter method for incoming votes and justifications. It is
applied when there is an existing session. If there is no session,
incoming message goes through validation and then we try to process it.
We call acceptInterval() and if meanwhile a beefy session is created, we
would have an existing one and will determine the round action. If there
is still no session, we invalidate the message.
- Enhanced checks in handling votes and justifications in
BeefyMessageHandler class.

Fixes #866 

### Checklist:

- [x] I have read the [contributing
guidelines](https://github.com/LimeChain/Fruzhin/blob/dev/CONTRIBUTING.md).
- [x] My PR title matches the [Conventional Commits
spec](https://www.conventionalcommits.org/).
- [ ] My change requires a change to the documentation.
- [ ]  I have updated the documentation accordingly.
- [ ]  I have added tests to cover my changes.
- [x] All new and existing tests passed.

---------

Co-authored-by: Hristiyan Mitov <hristiyan.mitov@limechain.tech>