feat: Improve neighbor message sending logic (#879)
# Description

### Workflow
Neighbor messages are broadcast:
- On block finalization;
- When the authority set changes;
- When switching to a new round;
- And periodically, every 5 minutes.

**Finalization of block**
We can finalize a block from a commit message and when we
`Attempt-to-Finalize`.

**Switch a setID**
When we handle authority changes from incoming blocks and from produced
blocks (when node is part of babe producers).

**Switch a round**
When we receive a commit message, that implies that the next round has
started, we should broadcast is(and maybe if we are active voters we
should start a new round, with number from the commit message). The
other scenario is when we play a new grandpa round(when node is active
authority).

### Changes in PR
- Renamed `sendMessagesToPeers()` method to
`sendNeighborMessageToPeers()` - it is more descriptive this way;
- Trigger neighbor message broadcast when a commit is received and
during `attemptToFinalize()`;
- Trigger neighbor message on starting of new authority set;
- Added setid, roundnum and lastfinalized to peerInfo with corresponding
update method, used on receiving of neighbor message;
- Added checks for peer set id and round num on receiving neighbor
messages with //TODOs for future rating trackings;
- Added check when broadcasting neighbor message - if peer set id is
same as ours and peer is Light client, do not broadcast.

Fixes #708 >

## Checklist:
- [x] I have read the [contributing
guidelines](https://github.com/LimeChain/Fruzhin/blob/dev/CONTRIBUTING.md).
- [x] My PR title matches the [Conventional Commits
spec](https://www.conventionalcommits.org/).
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.

---------

Co-authored-by: Hristiyan Mitov <hristiyan.mitov@limechain.tech>