feat: exchangeable runtime memory (#932)
# Issue

The block import was failing on block #29231. It ended up being an issue
regarding not properly handling updated runtime code.

# Solution

Implement the functionality to swap `DiskTrieAccessor` for the same
`Context`. This proves useful when having to keep in memory the changes
of each executed block before it is finalized. This is achieved via the
`TrieAccessorStorage` class. This was required in order to implement
`CodeChangeChecker`. The mention class has the responsibility to check
for changes between the current and saved runtime code. In case it finds
a changes it triggers the creation of a new runtime instance.

Fixes #493 

## Checklist:
- [X] I have read the [contributing
guidelines](https://github.com/LimeChain/Fruzhin/blob/dev/CONTRIBUTING.md).
- [X] My PR title matches the [Conventional Commits
spec](https://www.conventionalcommits.org/).
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [X] All new and existing tests passed.

---------

Co-authored-by: Viktor Todorov <viktor.todorov@limechain.tech>